Cocoa: move initial expose to drawRect
Move the initial expose event forward from setVisible
to drawRect. The expose event should now be sent
closer to the time when the window is actually visible,
but still early enough to give us a perfect first
frame.

Add a flushWindowSystemEvents() call to the end of
exposeWindow(): We need to make sure Qt processes
the expose and draws before drawRect returns.

Remove the flushWindowSystemEvents() call from setVisible(),
which was troublesome since we were calling it in
response to an event from Qt (as opposed to an event
from the OS)

Make the setFrame call in propagateSizeHints() not
send a paint event - we don't want to send the first
repaint here.

Change-Id: I5f2e4ec802dd1850317a9b9c45e48ca909827234