Make OpenGL Work in "non-requested" layer mode
The view may be in layer mode if a parent view requested
a layer. makeBackingLayer will then be called for
the view even if it did not call setWantsLayer:YES

Set m_layerMode based on whether makeBackingLayer was
called instead of based on whether a layer was requested
for the view.