Add QCocoaWindowFunctions::getNSView()
This function gets the NSView instance for a QWindow
instance and also transfers ownership of the QWindow
instance to the view. This means that the NSView will
delete the QWindow, and that application code must
not.

This is useful for cases where we want the lifetime
of the QWindow to follow that of the NSView, for example
if Qt is used to implement a plugin which provides
a NSView to a host application, where the host application
controls the lifetime of the view.

Change-Id: Ie059f44860d3a5966ad306dcefe7a97bb83f2b87